[yiimp-pool-status]: Display btcmhday in integer format

+++ yiimp-pool-status/main.go
@@ -76,8 +76,8 @@ func main() {
 		if factor == 0 {
 			factor = algo_mBTC_factor(key)
 		}
-		btcmhday := status[key].ActualLast24h * 1e5 / factor
-		fmt.Printf("\"%s\" \"yiimpstatus.%s.btcmhday[%s]\" \"%f\"\n", hostname, poolkey, element["ALGO"], btcmhday)
+		btcmhday := int64(status[key].ActualLast24h * 1e5 / factor)
+		fmt.Printf("\"%s\" \"yiimpstatus.%s.btcmhday[%s]\" \"%d\"\n", hostname, poolkey, element["ALGO"], btcmhday)
 		btctotal := status[key].Hashrate24h * status[key].ActualLast24h / factor / 1e9
 		fmt.Printf("\"%s\" \"yiimpstatus.%s.btctotal[%s]\" \"%f\"\n", hostname, poolkey, element["ALGO"], btctotal)
